refactor: combine core icon button components
Combine them into existing and separate file.
